Renamed factory and removed legacy code
