+++ contracts/app/andromeda-adodb/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "andromeda-factory"
-version = "0.1.0"
+name = "andromeda-adodb"
+version = "0.1.1"
 authors = ["Connor Barr <crnbarr@gmail.com>"]
 edition = "2018"
