+++ contracts/app/andromeda-adodb/examples/schema.rs
@@ -1,4 +1,4 @@
-use andromeda_app::factory::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use andromeda_app::adodb::{ExecuteMsg, InstantiateMsg, QueryMsg};
 use cosmwasm_schema::write_api;
 
 fn main() {
