+++ contracts/app/andromeda-adodb/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod contract;
-mod reply;
 mod state;
 
 #[cfg(test)]
