+++ packages/ado-base/src/instantiate.rs
@@ -40,7 +40,7 @@ impl<'a> ADOContract<'a> {
         name: &str,
     ) -> Result<u64, ContractError> {
         // Do we want to cache the factory address?
-        let factory_address = self.get_address_from_primitive(storage, querier, "factory")?;
+        let factory_address = self.get_address_from_primitive(storage, querier, "adodb")?;
         let code_id: u64 = query_get(Some(encode_binary(&name)?), factory_address, querier)?;
         Ok(code_id)
     }
