+++ packages/andromeda-app/src/adodb.rs
@@ -1,0 +1,24 @@
+use common::ado_base::{AndromedaMsg, AndromedaQuery};
+use cosmwasm_schema::{cw_serde, QueryResponses};
+
+#[cw_serde]
+pub struct InstantiateMsg {}
+
+#[cw_serde]
+pub enum ExecuteMsg {
+    AndrReceive(AndromedaMsg),
+    UpdateCodeId { code_id_key: String, code_id: u64 },
+}
+
+#[cw_serde]
+pub struct MigrateMsg {}
+
+#[cw_serde]
+#[derive(QueryResponses)]
+pub enum QueryMsg {
+    #[returns(AndromedaQuery)]
+    AndrQuery(AndromedaQuery),
+    /// All code IDs for Andromeda contracts
+    #[returns(u64)]
+    CodeId { key: String },
+}
