+++ packages/andromeda-app/src/lib.rs
@@ -1,2 +1,2 @@
+pub mod adodb;
 pub mod app;
-pub mod factory;
